Move getImageData to lib
