+++ lib/jsonld-schemas.js
@@ -1,24 +1,5 @@
 const _ = require('lodash');
-
-function getImageData(imageRelationship) {
-  if (!imageRelationship?.[0]) {
-    return null;
-  }
-
-  const img = imageRelationship[0];
-  const attachment = img.attachment;
-
-  if (!attachment?._urls) {
-    return null;
-  }
-
-  return {
-    url: attachment._urls.original || attachment._urls.full,
-    alt: img.alt || attachment.title || '',
-    width: attachment.width,
-    height: attachment.height
-  };
-}
+const { getImageData } = require('./utils');
 class JsonLdSchemaHandler {
   constructor() {
     this.schemas = {
@@ -73,7 +54,7 @@ class JsonLdSchemaHandler {
 
     // Add ItemList for listing pages if items are present and toggle allows
     const items = this.getListingItems(data);
-    const isDetailDoc = !!piece; // heuristic: piece detail pages shouldn’t emit ItemList
+    const isDetailDoc = !!piece; // heuristic: piece detail pages shouldn't emit ItemList
     const includeItemList = !isDetailDoc && items.length > 0 && (
       document?.seoIncludeItemList === true ||
       (document?.seoIncludeItemList === undefined && document?.seoJsonLdType === 'CollectionPage')
@@ -208,8 +189,6 @@ class JsonLdSchemaHandler {
     }
 
     if (data.req?.locale) schema.inLanguage = data.req.locale;
-    if (baseUrl) schema.publisher = { '@id': `${baseUrl}/#org` };
-
 
     // Only add description if it exists
     if (document.seoDescription?.trim() || document.excerpt?.trim()) {
@@ -234,31 +213,23 @@ class JsonLdSchemaHandler {
 
     // Add publisher from global org data
     const global = data.global;
-    // Add featured/publisher logo if available (guard for not-joined relations)
-    const publisherLogo = global?.seoJsonLdOrganization?._logo
-      ? getImageData(global.seoJsonLdOrganization._logo)
-      : null;
-    if (publisherLogo) {
-      // ensure publisher exists before assigning logo
-      if (!schema.publisher) {
-        schema.publisher = { '@type': 'Organization' };
-      }
-      schema.publisher.logo = {
-        '@type': 'ImageObject',
-        url: publisherLogo.url
-      };
-    }
-
     if (global?.seoJsonLdOrganization?.name?.trim()) {
       schema.publisher = {
         '@type': 'Organization',
         'name': global.seoJsonLdOrganization.name
       };
-
-      if (global.seoJsonLdOrganization._logo?.length && global.seoJsonLdOrganization._logo[0]._urls) {
+      
+      // Add @id reference if we have baseUrl (links to main Organization in graph)
+      if (baseUrl) {
+        schema.publisher['@id'] = `${baseUrl}/#org`;
+      }
+      
+      // Add logo if available
+      const publisherLogo = getImageData(global.seoJsonLdOrganization._logo);
+      if (publisherLogo) {
         schema.publisher.logo = {
           '@type': 'ImageObject',
-          'url': global.seoJsonLdOrganization._logo[0]._urls.original
+          'url': publisherLogo.url
         };
       }
     }
@@ -568,7 +539,7 @@ class JsonLdSchemaHandler {
         return new URL(abs).origin;
       } catch (e) { }
     }
-
+    
     return null;
   }
 
@@ -690,19 +661,18 @@ class JsonLdSchemaHandler {
     const thumbnail = getImageData(video._thumbnail) || getImageData(document._featuredImage);
     if (thumbnail) {
       schema.thumbnailUrl = thumbnail.url;
-
-      // Video URL or embed URL
-      if (video.contentUrl?.trim()) {
-        schema.contentUrl = video.contentUrl;
-      }
-
-      if (video.embedUrl?.trim()) {
-        schema.embedUrl = video.embedUrl;
-      }
-
     }
 
-      return schema;
+    // Video URL or embed URL
+    if (video.contentUrl?.trim()) {
+      schema.contentUrl = video.contentUrl;
+    }
+
+    if (video.embedUrl?.trim()) {
+      schema.embedUrl = video.embedUrl;
+    }
+
+    return schema;
   }
 }
